refactor(hydro_lang): share compilation logic between Maelstrom and sim targets
The Maelstrom deployment target previously compiled its generated binary with a
bespoke, statically-linked `cargo build --example` invocation that recompiled the
whole dependency tree on every run and had no concurrency management. The sim
target already had a parallel-compilation pipeline (per-job target dirs with
symlinked shared artifacts, plus a prebuilt dylib of the dependencies). This
change makes Maelstrom reuse that same pipeline.

- Extract the concurrent-build machinery out of sim's `compile_sim` into a shared
  `compile_trybuild_example(ExampleBuildConfig)` in `compile/trybuild/generate.rs`,
  parameterized over the runtime feature, example name, crate type (`cdylib` for
  sim vs. an executable for Maelstrom), the `TRYBUILD_LIB_NAME` indirection, and
  whether fuzzing is allowed.
- Make `LinkingMode::Dynamic`, `TrybuildConfig::linking_mode`, and the dynamic
  examples-dir path available under the `maelstrom` feature (previously
  `deploy`-only).
- `compile_sim` is now a thin wrapper over the shared function.
- Maelstrom instantiates with `LinkingMode::Dynamic`, stores the full
  `TrybuildConfig`, and `build()` uses the shared builder, persisting the built
  executable via `TempPath::keep`.
- The `maelstrom` feature now depends on `hydro_concurrent_cargo` and
  `cargo_metadata`.

Fixes #2985.

PR: #2998
